requirements.txt
streamlit
numpy
Pillow
rembg
opencv-python-headless
